Fixed a few more typing-related issues for resource specs

+++ internal/server/jupyter/message.go
@@ -31,7 +31,7 @@ type KernelMessage interface {
 
 // ResourceSpec can be passed within a jupyterSessionReq when creating a new Session or Kernel.
 type ResourceSpec struct {
-	Cpu float64 `json:"cpu"`    // In millicpus (1/1000th CPU core)
+	Cpu int     `json:"cpu"`    // In millicpus (1/1000th CPU core)
 	Mem float64 `json:"memory"` // In MB
 	Gpu int     `json:"gpu"`
 }
